requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
google-generativeai==0.3.2
pandas==2.2.0
jinja2==3.1.3
reportlab==4.0.9
python-dotenv==1.0.0
